move db helpers to `/lib` folder
