+++ lib/msfdb_helpers/db_interface.rb
